third_party/blink/web_tests/external/wpt/speculation-rules/prefetch/tentative/cookie-indices.https_include=changed-expected.txt

This is a testharness.js-based test.
[FAIL] Cookie-Indices should prevent a prefetch from being used if the cookie has changed.
  assert_equals: expected (undefined) undefined but got (string) "prefetch"
Harness: the test ran to completion.
